Separate the kernel layers, and stop duplicating the pipeline and the reference
All from Bruno's review.

**Layering.** Kernel-layer modules in this repo import no torch -- both neighbours in
`_gluon_kernels/gfx950/attention/` have zero torch references, and so does everything under
`_triton_kernels/attention/`. Ours was the exception. The two `@gluon.jit` kernels stay in
`_gluon_kernels/gfx950/attention/`, the two `@triton.jit` kernels move to
`_triton_kernels/attention/`, and the five launchers plus the CSR build and the sink gradient
move up to `attention/sparse_attention_dsv4_bwd.py`. Both kernel modules are now torch-free with
no public functions, matching their neighbours, so the kernels can be reached without pulling in
torch.

The Triton path is `_triton_kernels/attention/`, not `_triton_kernels/gfx950/attention/` as the
comment suggested: that tree has no arch level. `pa_decode_sparse`, `fp8_mqa_logits` and `mla`
all sit arch-namespaced under `_gluon_kernels/` and flat under `_triton_kernels/`.

**Pipeline duplication.** The benchmark was restating the production orchestration -- phase
order, tile widths, workspace, and a copy of the sink-gradient expression. It now shares the
real thing: `plan_bwd()` validates and allocates, `bwd_phases()` yields `(name, thunk)` in
execution order, `sparse_mla_bwd_dsv4` runs them, and the benchmark times them one at a time.
The measurement is of the phases the op actually runs, and there is nothing left to drift. The
sink gradient lives once, on the plan object.

**Reference duplication.** The fp32 reference forward existed in both the test and the
benchmark. It moves to `aiter.test_mha_common.sparse_mla_dsv4_ref`, beside the other reference
implementations, and returns `(o, lse)` from one differentiable pass so backward values come
from autograd through it. Both callers use it; the test file loses 46 lines.

**Benchmark harness.** `triton.testing.do_bench` replaces the hand-rolled event loop -- 55 of
the 67 benchmarks here already use it. Help text added to `--cfgs` and `--skip-correctness`.

do_bench flushes L2 between repetitions, so the reported numbers move: 3.271 ms / 420 TFLOPS
end-to-end at T=4096 H=128 topk=512, against 3.229 / 426 with the previous cache-warm loop. The
per-phase figures move further for the same reason and now sum slightly above the end-to-end
number. The PR description carries the new tables.

8/8 tests pass on gfx950 (MI355X).
