+++ aiter/ops/triton/_gluon_kernels/gfx950/attention/sparse_attention_dsv4_bwd.py
@@ -1,14 +1,11 @@
 # SPDX-License-Identifier: MIT
 # Copyright (C) 2024-2026, Advanced Micro Devices, Inc. All rights reserved.
 
-"""Kernels for the DeepSeek-V4 sparse-MLA training BACKWARD (gfx950 / CDNA4).
+"""Gluon kernels for the DeepSeek-V4 sparse-MLA training BACKWARD (gfx950 / CDNA4).
 
-All operate on the official V4 form (``K == V == kv``, one dense 512-wide tensor, RoPE already
-applied in place caller-side, scale ``1/sqrt(512)``, ``attn_sink`` in the softmax denominator
-only, ``topk == -1`` masked). The two MFMA phases are Gluon; the two memory-bound phases are
-plain Triton and live alongside them rather than under ``_triton_kernels/`` because there is no
-Triton implementation of this backward to fall back to -- they are parts of this kernel, not an
-alternative to it.
+The two MFMA phases. Both operate on the official V4 form: ``K == V == kv`` as one dense
+512-wide tensor, RoPE already applied in place caller-side, scale ``1/sqrt(512)``, ``attn_sink``
+in the softmax denominator only, ``topk == -1`` masked.
 
 ``_dq_v4_kernel``
     Per (query token, head block): ``S = Q@kv^T``, ``P = exp(S - lse)``, ``dP = dO@kv^T``,
@@ -22,21 +19,10 @@ alternative to it.
     are transposed once into registers and D is split across ``grid.y``, which is what keeps them
     read once instead of ``topk/TILE_K`` times.
 
-``_delta_v4_kernel``
-    ``delta = rowsum(O * dO)`` -- the standard flash-attention "o_dot_do" preamble. Streams the
-    bf16 inputs and accumulates in fp32, so it moves exactly the working set.
-
-``_bwd_dkv_gather_acc_v4`` + ``build_inverted_topk``
-    Reduce ``interm[t, slot, :]`` into ``dkv[kv_row, :]`` over the top-k mapping. The scatter is
-    inverted into a CSR gather (each output KV row collects its own contributors), so no atomics
-    are needed. ``BLOCK_E`` entries are carried per loop iteration, which both widens the load
-    and cuts the trip count on the long runs a realistic top-k produces.
-
-Public entry: ``aiter.ops.triton.attention.sparse_attention_dsv4_bwd.sparse_mla_bwd_dsv4``.
+Launchers live in ``aiter.ops.triton.attention.sparse_attention_dsv4_bwd``; this module stays
+free of torch so the kernels can be called without it.
 """
 
-import torch
-import triton
 import triton.language as tl
 from triton.experimental import gluon
 from triton.experimental.gluon import language as gl
@@ -364,59 +350,6 @@ def _dq_v4_kernel(
     )
 
 
-def sparse_mla_bwd_dq(
-    q,
-    kv,
-    do,
-    topk,
-    lse,
-    delta,
-    dq,
-    chunk_dS,
-    chunk_P,
-    scale,
-    r_start,
-    R_CHUNK,
-    BLOCK_H=64,
-    TILE_K=32,
-    is_first_chunk=True,
-):
-    """Launch the dQ kernel for one rank chunk. Writes ``dq`` (RMW when not the first chunk)
-    plus this chunk's ``chunk_dS`` / ``chunk_P``."""
-    T, H, D = q.shape
-    _dq_v4_kernel[(T, triton.cdiv(H, BLOCK_H))](
-        q,
-        kv,
-        do,
-        topk,
-        lse,
-        delta,
-        dq,
-        chunk_dS,
-        chunk_P,
-        q.stride(0),
-        q.stride(1),
-        kv.stride(0),
-        do.stride(0),
-        do.stride(1),
-        dq.stride(0),
-        dq.stride(1),
-        topk.stride(0),
-        chunk_dS.stride(0),
-        chunk_dS.stride(1),
-        scale,
-        H,
-        r_start,
-        R_CHUNK=R_CHUNK,
-        BLOCK_H=BLOCK_H,
-        TILE_K=TILE_K,
-        D=D,
-        IS_FIRST_CHUNK=is_first_chunk,
-        num_warps=4,
-        waves_per_eu=1,
-    )
-
-
 _dkv_interm_v4_kernel_repr = make_kernel_repr(
     "_dkv_interm_v4_kernel",
     [
@@ -598,219 +531,3 @@ def _dkv_interm_v4_kernel(
             ptr=Interm_ptr,
             offsets=interm_offs.to(tl.int32),
         )
-
-
-def sparse_mla_bwd_dkv_interm_v4(
-    q,
-    do,
-    chunk_dS,
-    chunk_P,
-    R_CHUNK,
-    BD=256,
-    TILE_K=128,
-    MFMA_K=32,
-    DUAL_STAGE=1,
-    H_POW2=None,
-    num_warps=4,
-    interm=None,
-):
-    """V4 dKV-interm, Q/dO read once. Returns interm [T, R_CHUNK, D] bf16.
-
-    ``BD`` splits D across ``grid.y``; dS/P are re-read once per D block, so a larger BD moves
-    less of them. ``MFMA_K=32`` is the CDNA4 16x16x32 depth.
-    """
-    T, H, D = q.shape
-    assert R_CHUNK % TILE_K == 0
-    assert D % BD == 0
-    h_pow2 = H_POW2 or triton.next_power_of_2(H)
-    if interm is None:
-        interm = torch.empty(T, R_CHUNK, D, dtype=torch.bfloat16, device=q.device)
-    _dkv_interm_v4_kernel[(T, D // BD)](
-        q,
-        do,
-        chunk_dS,
-        chunk_P,
-        interm,
-        q.stride(0),
-        q.stride(1),
-        do.stride(0),
-        do.stride(1),
-        chunk_dS.stride(0),
-        chunk_dS.stride(1),
-        interm.stride(0),
-        interm.stride(1),
-        H,
-        R_CHUNK=R_CHUNK,
-        TILE_K=TILE_K,
-        NH=h_pow2,
-        BD=BD,
-        D=D,
-        MFMA_K=MFMA_K,
-        DUAL_STAGE=DUAL_STAGE,
-        num_warps=num_warps,
-    )
-    return interm
-
-
-_delta_v4_kernel_repr = make_kernel_repr(
-    "_delta_v4_kernel",
-    [
-        "D",
-        "BLOCK_R",
-    ],
-)
-
-
-@triton.jit(repr=_delta_v4_kernel_repr)
-def _delta_v4_kernel(
-    O_ptr,  # [n_rows, D] bf16   (rows = T*H, contiguous)
-    dO_ptr,  # [n_rows, D] bf16
-    Delta_ptr,  # [n_rows]    fp32
-    n_rows,
-    D: tl.constexpr,
-    BLOCK_R: tl.constexpr,
-):
-    """Grid (cdiv(n_rows, BLOCK_R),) — each program reduces BLOCK_R rows of width D."""
-    pid = tl.program_id(0)
-    rows = pid * BLOCK_R + tl.arange(0, BLOCK_R)
-    mask = rows < n_rows
-    offs = rows.to(tl.int64)[:, None] * D + tl.arange(0, D)[None, :]
-    o = tl.load(O_ptr + offs, mask=mask[:, None], other=0.0).to(tl.float32)
-    d = tl.load(dO_ptr + offs, mask=mask[:, None], other=0.0).to(tl.float32)
-    tl.store(Delta_ptr + rows, tl.sum(o * d, axis=1), mask=mask)
-
-
-def delta_v4(o, do, out=None, BLOCK_R=8, num_warps=8):
-    # BLOCK_R=8 keeps each lane loading >= 8 bf16, i.e. a dwordx4; narrower blocks drop to a
-    # dword and the kernel loses most of its bandwidth.
-    """o[T,H,D] bf16, do[T,H,D] bf16 -> delta[T,H] fp32 = sum_d o*do.
-
-    ``do`` must already be the D-wide (lora) slice, contiguous — same contract as the dQ kernel.
-    """
-    assert o.shape == do.shape and o.is_contiguous() and do.is_contiguous()
-    T, H, D = o.shape
-    n_rows = T * H
-    if out is None:
-        out = torch.empty(T, H, dtype=torch.float32, device=o.device)
-    _delta_v4_kernel[(triton.cdiv(n_rows, BLOCK_R),)](
-        o,
-        do,
-        out,
-        n_rows,
-        D=D,
-        BLOCK_R=BLOCK_R,
-        num_warps=num_warps,
-    )
-    return out
-
-
-_bwd_dkv_gather_acc_v4_repr = make_kernel_repr(
-    "_bwd_dkv_gather_acc_v4",
-    [
-        "D",
-        "BLOCK_E",
-        "ACCUMULATE",
-    ],
-)
-
-
-@triton.jit(repr=_bwd_dkv_gather_acc_v4_repr)
-def _bwd_dkv_gather_acc_v4(
-    Interm_ptr,  # [T, R_CHUNK, D] bf16, flat [T*R_CHUNK, D]
-    InvPtr_ptr,  # [num_kv+1] int32 — CSR row pointers
-    InvData_ptr,  # [valid] int32 — encoded q*R_CHUNK+local_r, sorted by KV token
-    dKV_acc_ptr,  # [num_kv, D] fp32 — accumulator
-    stride_interm_r: tl.int64,
-    stride_acc_t: tl.int64,
-    D: tl.constexpr,
-    BLOCK_E: tl.constexpr,
-    ACCUMULATE: tl.constexpr,
-):
-    """Grid (num_kv,) — one CTA per KV token, BLOCK_E CSR entries in flight.
-
-    ``BLOCK_E`` entries are carried per iteration, which the gather needs for two reasons:
-
-      * **load width.** A bare ``tl.arange(0, D)`` block over 256 threads is 2 bf16 = 4 B per
-        lane -- a dword. The [BLOCK_E, D] block gives ``BLOCK_E*D/threads`` elements per lane,
-        so the loads become dwordx4. The gather is issue-bound, so this dominates.
-      * **trip count.** ``tl.sum`` folds the entry axis, so the run is consumed BLOCK_E at a
-        time. A realistic top-k gives run lengths up to ~3000 on the pool rows.
-
-    ``ACCUMULATE=False`` writes the destination instead of reading it back first. The caller
-    uses it for the first chunk, where the accumulator is still zero.
-    """
-    k = tl.program_id(0)
-    offs_d = tl.arange(0, D)
-    offs_e = tl.arange(0, BLOCK_E)
-    start = tl.load(InvPtr_ptr + k)
-    end = tl.load(InvPtr_ptr + k + 1)
-    acc_base = k.to(tl.int64) * stride_acc_t
-
-    if ACCUMULATE:
-        acc = tl.load(dKV_acc_ptr + acc_base + offs_d).to(tl.float32)
-    else:
-        acc = tl.zeros([D], dtype=tl.float32)
-
-    for i0 in range(start, end, BLOCK_E):
-        idx = i0 + offs_e
-        m = idx < end
-        entry = tl.load(InvData_ptr + idx, mask=m, other=0).to(tl.int64)
-        vals = tl.load(
-            Interm_ptr + entry[:, None] * stride_interm_r + offs_d[None, :],
-            mask=m[:, None],
-            other=0.0,
-        )
-        acc += tl.sum(vals.to(tl.float32), axis=0)
-
-    tl.store(dKV_acc_ptr + acc_base + offs_d, acc)
-
-
-def build_inverted_topk(topk_indices_slice, num_kv):
-    """CSR inverted index over ``num_kv`` KV rows.
-
-    One stable sort yields both the permutation (``inv_data``) and the sorted keys;
-    ``inv_ptr[k] = searchsorted(sorted, k, 'left')`` = the number of entries with value < k.
-    Invalid (-1) entries sort to the front, so ``inv_ptr[0]`` starts past them and they are
-    never visited.
-
-    The sort key is narrowed to int16 when ``num_kv`` fits, which is what keeps the radix sort
-    to two byte-passes.
-
-    Returns ``inv_ptr[num_kv+1]`` int32, ``inv_data[T*R]`` int32.
-    """
-    # row_ids is the searchsorted query: [0 .. num_kv], one per KV row plus the end sentinel.
-    # Its dtype must match `keys` -- searchsorted is built per branch for that reason, not by
-    # accident.
-    flat_kv = topk_indices_slice.reshape(-1)  # [T*R] int32; -1 = invalid
-    if num_kv < 32767:  # int16 range, -1 included
-        keys = flat_kv.to(torch.int16)
-        row_ids = torch.arange(num_kv + 1, device=flat_kv.device, dtype=torch.int16)
-    else:
-        keys = flat_kv.to(torch.int32)
-        row_ids = torch.arange(num_kv + 1, device=flat_kv.device, dtype=torch.int32)
-    sorted_vals, inv_data = torch.sort(keys, stable=True)
-    inv_ptr = torch.searchsorted(sorted_vals, row_ids).to(torch.int32)
-    return inv_ptr, inv_data.to(torch.int32)
-
-
-def dkv_gather_acc(
-    interm, inv_ptr, inv_data, dkv_acc, BLOCK_E=64, num_warps=8, accumulate=True
-):
-    """interm[T,R,D] bf16 -> dkv_acc[num_kv,D] fp32 via the entry-blocked CSR gather.
-
-    Grid is ``num_kv`` (from ``dkv_acc``), not ``T``, so a compressed-pool KV works.
-    """
-    _, _, D = interm.shape
-    num_kv = dkv_acc.shape[0]
-    _bwd_dkv_gather_acc_v4[(num_kv,)](
-        interm,
-        inv_ptr,
-        inv_data,
-        dkv_acc,
-        interm.stride(1),
-        dkv_acc.stride(0),
-        D=D,
-        BLOCK_E=BLOCK_E,
-        ACCUMULATE=accumulate,
-        num_warps=num_warps,
-    )
