+++ aiter/test_mha_common.py
@@ -678,3 +678,45 @@ def attention_ref_with_tol(q, k, v, do, is_fp8=False, **kwargs):
     bwd_tols = [_tol(dq, dq_pt), _tol(dk, dk_pt), _tol(dv, dv_pt)]
 
     return out, (dq, dk, dv), fwd_tol, bwd_tols
+
+
+def sparse_mla_dsv4_ref(q, kv, topk_indices, attn_sink=None, scale=None):
+    """fp32 reference forward for the DeepSeek-V4 sparse MLA attention.
+
+    The official V4 form: shared-KV GQA where ``K == V == kv`` is one dense ``head_dim``-wide
+    tensor, RoPE already applied in place caller-side, ``attn_sink`` folded into the softmax
+    denominator only, and ``topk_indices == -1`` masked out.
+
+    Differentiable in ``q`` / ``kv`` / ``attn_sink``, so a backward reference is just autograd
+    through this. It is a per-token python loop, so keep the shapes small.
+
+    Args:
+        q:            [T, H, D] float32, requires_grad for a backward reference
+        kv:           [num_kv, D] float32, ``num_kv >= T``; rows ``T..`` are the compressed pool
+        topk_indices: [T, TOPK] int, -1 marks an invalid slot
+        attn_sink:    [H] float32 or None
+        scale:        softmax scale, defaults to ``1/sqrt(D)``
+
+    Returns:
+        ``(o, lse)`` -- ``o`` [T, H, D] float32 carrying grad, and ``lse`` [T, H] float32
+        detached and sink-inclusive, which is the convention the backward kernel expects.
+    """
+    if scale is None:
+        scale = 1.0 / (q.shape[-1] ** 0.5)
+    outs, lses = [], []
+    for t in range(q.shape[0]):
+        idx = topk_indices[t].long()
+        valid = idx != -1
+        k = kv[idx.clamp(min=0)]
+        k = torch.where(valid[:, None], k, torch.zeros_like(k))
+        s = (q[t] @ k.t()) * scale
+        s = torch.where(valid[None, :], s, torch.full_like(s, float("-inf")))
+        row_max = s.max(dim=1).values
+        m = row_max if attn_sink is None else torch.maximum(row_max, attn_sink)
+        p = torch.where(valid[None, :], torch.exp(s - m[:, None]), torch.zeros_like(s))
+        denom = p.sum(dim=1)
+        if attn_sink is not None:
+            denom = denom + torch.exp(attn_sink - m)
+        outs.append((p @ k) / denom[:, None])
+        lses.append((m + torch.log(denom)).detach())
+    return torch.stack(outs, dim=0), torch.stack(lses, dim=0)
