+++ op_tests/triton_tests/attention/test_sparse_attention_dsv4_bwd.py
@@ -3,10 +3,11 @@
 
 """Correctness for the DSv4 sparse-MLA training backward against torch autograd.
 
-The reference is a differentiable fp32 re-implementation of the V4 forward, differentiated by
-autograd -- an independent path from the kernels, not a re-expression of them. It is a per-token
-python loop, so the shapes here are deliberately small; the kernels are exercised at production
-shapes (T=4096, H=128, topk=512) out of tree.
+The reference is ``aiter.test_mha_common.sparse_mla_dsv4_ref``, a differentiable fp32
+re-implementation of the V4 forward -- an independent path from the kernels, not a re-expression
+of them. Backward values come from autograd through it. It is a per-token python loop, so the
+shapes here are deliberately small; the op benchmark exercises the kernels at production shapes
+and shares the same reference for its correctness gate.
 """
 
 import pytest
@@ -14,6 +15,7 @@ import torch
 
 from aiter.ops.triton.attention.sparse_attention_dsv4_bwd import sparse_mla_bwd_dsv4
 from aiter.ops.triton.utils._triton import arch_info
+from aiter.test_mha_common import sparse_mla_dsv4_ref
 
 D = 512
 COS_TOL = 0.999
@@ -22,55 +24,6 @@ pytestmark = pytest.mark.skipif(
     arch_info.get_arch() != "gfx950",
     reason="DSv4 sparse-MLA backward is gfx950 (CDNA4) only",
 )
-
-
-def _ref_fwd_diff(q, kv, attn_sink, topk, scale):
-    """Differentiable fp32 V4 forward. q/kv/attn_sink require grad. Returns O [T, H, D]."""
-    outs = []
-    for t in range(q.shape[0]):
-        idx = topk[t].long()
-        valid = idx != -1
-        k = kv[idx.clamp(min=0)]
-        k = torch.where(valid[:, None], k, torch.zeros_like(k))
-        s = (q[t] @ k.t()) * scale
-        s = torch.where(valid[None, :], s, torch.full_like(s, float("-inf")))
-        if attn_sink is None:
-            m = s.max(dim=1).values
-            p = torch.where(
-                valid[None, :], torch.exp(s - m[:, None]), torch.zeros_like(s)
-            )
-            denom = p.sum(dim=1)
-        else:
-            m = torch.maximum(s.max(dim=1).values, attn_sink)
-            p = torch.where(
-                valid[None, :], torch.exp(s - m[:, None]), torch.zeros_like(s)
-            )
-            denom = p.sum(dim=1) + torch.exp(attn_sink - m)
-        outs.append((p @ k) / denom[:, None])
-    return torch.stack(outs, dim=0)
-
-
-def _ref_fwd(q, kv, attn_sink, topk, scale):
-    """Non-differentiable forward giving the kernel inputs O (bf16) and sink-inclusive lse."""
-    with torch.no_grad():
-        o = _ref_fwd_diff(q.float(), kv.float(), attn_sink, topk, scale)
-        lse = torch.empty(q.shape[0], q.shape[1], device=q.device, dtype=torch.float32)
-        for t in range(q.shape[0]):
-            idx = topk[t].long()
-            valid = idx != -1
-            k = kv.float()[idx.clamp(min=0)]
-            k = torch.where(valid[:, None], k, torch.zeros_like(k))
-            s = (q.float()[t] @ k.t()) * scale
-            s = torch.where(valid[None, :], s, torch.full_like(s, float("-inf")))
-            m = s.max(dim=1).values
-            p = torch.where(
-                valid[None, :], torch.exp(s - m[:, None]), torch.zeros_like(s)
-            )
-            denom = p.sum(dim=1)
-            if attn_sink is not None:
-                denom = denom + torch.exp(attn_sink - m)
-            lse[t] = m + torch.log(denom)
-    return o.to(torch.bfloat16).contiguous(), lse
 
 
 def _cos(a, b):
@@ -104,16 +57,17 @@ def test_sparse_mla_bwd_dsv4(T, H, topk, npool, has_sink, r_chunk):
     invalid = torch.rand(T, topk, device=dev) < 0.1
     indices = torch.where(invalid, torch.full_like(indices, -1), indices).contiguous()
 
-    o, lse = _ref_fwd(q, kv, sink, indices, scale)
+    qg = q.float().clone().requires_grad_(True)
+    kvg = kv.float().clone().requires_grad_(True)
+    sg = sink.clone().requires_grad_(True) if has_sink else None
+    ref_o, lse = sparse_mla_dsv4_ref(qg, kvg, indices, sg, scale)
+    o = ref_o.detach().to(torch.bfloat16).contiguous()
 
     dq, dkv, d_sink = sparse_mla_bwd_dsv4(
         q, kv, do, o, lse, indices, attn_sink=sink, scale=scale, R_CHUNK=r_chunk
     )
 
-    qg = q.float().clone().requires_grad_(True)
-    kvg = kv.float().clone().requires_grad_(True)
-    sg = sink.clone().requires_grad_(True) if has_sink else None
-    _ref_fwd_diff(qg, kvg, sg, indices, scale).backward(do.float())
+    ref_o.backward(do.float())
 
     assert _cos(dq, qg.grad) > COS_TOL, f"dq cos {_cos(dq, qg.grad)}"
     assert _cos(dkv, kvg.grad) > COS_TOL, f"dkv cos {_cos(dkv, kvg.grad)}"
